+++ OctoCmd/Resources/Assets.xcassets/AppIcon.appiconset/icon_512x512@2x.png
