+++ OctoCmd/Resources/Assets.xcassets/Contents.json
