+++ OctoCmd/Resources/Info.plist
