+++ OctoCmd/Resources/OctoCmd.entitlements
